requirements.txt
pyqt5
python-nmap
fpdf
